addded product_sales column to table to keep track of total sales by item.

+++ bmazon_schema.sql
@@ -10,13 +10,13 @@ CREATE TABLE products(
     department_name VARCHAR(50) NOT NULL,
     price DECIMAL(10,2) NOT NULL,
     stock_quantity INT NULL,
-    product_sales INT NULL,
+    product_sales DECIMAL(10,2) NULL,
     PRIMARY KEY (item_id)
 );
 
 CREATE TABLE departments (
     department_id INT NOT NULL AUTO_INCREMENT,
     department_name VARCHAR(50) NOT NULL,
-    over_head_costs INT(10) NOT NULL,
+    over_head_costs DECIMAL(10,2) NOT NULL,
     PRIMARY KEY (department_id)
 );
